backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Config & validation
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Auth / JWT
python-jose[cryptography]==3.3.0
email-validator==2.2.0

# Payments
stripe==11.4.1

# Email
resend==2.7.0

# HTTP client (for external APIs)
httpx==0.28.0

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx  # also used as async test client
